Updated to use vs2017

+++ ObjGLUF/GLUFGui.cpp
@@ -7286,7 +7286,7 @@ void EditBox::UpdateRectsMultiline() noexcept
     NOEXCEPT_REGION_START
 
 
-    mScrollBar->SetRegion({ { mTextRegion.right - mSBWidth }, mSubRegions[0].top, mSubRegions[0].right, { mSubRegions[0].bottom } });
+    mScrollBar->SetRegion({ { mTextRegion.right - static_cast<long>(mSBWidth) }, mSubRegions[0].top, mSubRegions[0].right, { mSubRegions[0].bottom } });
     mScrollBar->SetPageSize(static_cast<int>(RectHeight(mTextRegion) / mDialog.GetFont(mElements[0].mFontIndex)->mLeading));
 
     mTextRegion.right -= 2 * mSBWidth;
